Move StreamConverters to javaapi
